+++ commands/app-sethiddenfile.js
@@ -1,12 +1,12 @@
-var chalk = require('chalk');
-var shell = require('shelljs');
-var fileExists = require('file-exists');
-var pathExists = require('path-exists');
-var Loading = require('./loading');
+let chalk = require('chalk');
+let shell = require('shelljs');
+let fileExists = require('file-exists');
+let pathExists = require('path-exists');
+let Loading = require('./loading');
 
-var SetupHiddenFiles = (function (isNewVersion, modulePath) {
+let SetupHiddenFiles = (function (isNewVersion, modulePath) {
 	
-	var _CONFIG = {
+	let _CONFIG = {
 		isNewVersion: isNewVersion,
 		corePath: 'webpack'
 	};
@@ -26,7 +26,7 @@ var SetupHiddenFiles = (function (isNewVersion, modulePath) {
 	
 	function yeah(fileName) {
 		
-		var _tempPath = process.env.PWD + '/.ciffi/';
+		let _tempPath = process.env.PWD + '/.ciffi/';
 		
 		pathExists(_tempPath).then(function (res) {
 			if (!res) {
@@ -34,10 +34,10 @@ var SetupHiddenFiles = (function (isNewVersion, modulePath) {
 			}
 		});
 		
-		var _tempFile = _tempPath + fileName;
-		var _resource = modulePath + '/lib/node_modules/ciffi/node_modules/ciffi-js-' + _CONFIG.corePath + '/resources/core/' + fileName;
-		var _projectRoot = process.env.PWD + '/';
-		var _projectFile = process.env.PWD + '/.' + fileName;
+		let _tempFile = _tempPath + fileName;
+		let _resource = modulePath + '/lib/node_modules/ciffi/node_modules/ciffi-js-' + _CONFIG.corePath + '/resources/core/' + fileName;
+		let _projectRoot = process.env.PWD + '/';
+		let _projectFile = process.env.PWD + '/.' + fileName;
 		
 		if (fileExists.sync(_projectFile)) {
 			console.log(chalk.red('File already exists: ' + _projectFile));