+++ commands/tempApp.js
@@ -1,18 +1,18 @@
-var shell = require('shelljs');
-var replace = require('replace-in-file');
-var pathExists = require('path-exists');
+let shell = require('shelljs');
+let replace = require('replace-in-file');
+let pathExists = require('path-exists');
 
-var TempApp = (function (isNewVersion, modulePath, callback) {
+let TempApp = (function (isNewVersion, modulePath, callback) {
 	
-	var _CONFIG = {
+	let _CONFIG = {
 		isNewVersion: isNewVersion,
 		callback: callback,
 		frontendPath: 'webpack'
 	};
 	
 	function TempApp() {
-		var _modulePath = modulePath + '/lib/node_modules/ciffi/node_modules/ciffi-js-' + _CONFIG.frontendPath + '/frontend/*';
-		var _tempPath = process.env.PWD + '/.ciffi/';
+		let _modulePath = modulePath + '/lib/node_modules/ciffi/node_modules/ciffi-js-' + _CONFIG.frontendPath + '/frontend/*';
+		let _tempPath = process.env.PWD + '/.ciffi/';
 		
 		pathExists(_tempPath).then(function (res) {
 			if (!res) {